organize input element types and implementation into dedicated files
