+++ crates/gpui_elements/src/input.rs
@@ -1,5 +1,15 @@
-mod element;
-pub use element::*;
-
+pub mod actions;
+mod colors;
 mod cursor;
-pub use cursor::*;
+mod element;
+mod history;
+mod paint;
+mod state;
+mod state_input_handler;
+pub(self) mod unicode;
+
+pub use colors::*;
+pub(self) use cursor::*;
+pub use element::*;
+pub(self) use history::*;
+pub use state::*;
